+++ changelog.md
@@ -1,6 +1,6 @@
 # Ruck changelog
 
-## Next
+## 6.0.0
 
 ### Major